fix: add and edit workspace member UI
