+++ webui/react/src/pages/WorkspaceDetails/RoleRenderer.module.scss
@@ -1,0 +1,3 @@
+.base {
+  min-width: 195px;
+}
