+++ webui/react/src/pages/WorkspaceDetails/WorkspaceMembers.module.scss
@@ -23,6 +23,3 @@
 .membersContainer {
   margin-top: 30px;
 }
-.selectContainer {
-  min-width: 195px;
-}
