+++ webui/react/src/pages/WorkspaceDetails/WorkspaceMembers.tsx
@@ -1,7 +1,5 @@
-import { Button, Dropdown, Menu, Select } from 'antd';
+import { Button, Dropdown, Menu } from 'antd';
 import { FilterDropdownProps } from 'antd/lib/table/interface';
-import { RawValueType } from 'rc-select/lib/BaseSelect';
-import { LabelInValueType } from 'rc-select/lib/Select';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
 import InteractiveTable, {
@@ -11,31 +9,19 @@ import InteractiveTable, {
 import { getFullPaginationConfig } from 'components/Table/Table';
 import TableFilterSearch from 'components/Table/TableFilterSearch';
 import Avatar from 'components/UserAvatar';
-import { useStore } from 'contexts/Store';
 import useFeature from 'hooks/useFeature';
 import { useFetchKnownRoles } from 'hooks/useFetch';
 import useModalWorkspaceRemoveMember from 'hooks/useModal/Workspace/useModalWorkspaceRemoveMember';
 import usePermissions from 'hooks/usePermissions';
 import useSettings, { UpdateSettings } from 'hooks/useSettings';
-import {
-  assignRolesToGroup,
-  assignRolesToUser,
-  removeRolesFromGroup,
-  removeRolesFromUser,
-} from 'services/api';
-import {
-  V1Group,
-  V1GroupDetails,
-  V1RoleAssignment,
-  V1RoleWithAssignments,
-} from 'services/api-ts-sdk';
+import { V1Group, V1GroupDetails, V1RoleWithAssignments } from 'services/api-ts-sdk';
 import { Size } from 'shared/components/Avatar';
 import Icon from 'shared/components/Icon/Icon';
 import { alphaNumericSorter } from 'shared/utils/sort';
 import { User, UserOrGroup, Workspace } from 'types';
-import handleError from 'utils/error';
-import { getIdFromUserOrGroup, getName, isUser } from 'utils/user';
+import { getAssignedRole, getIdFromUserOrGroup, getName, isUser } from 'utils/user';
 
+import RoleRenderer from './RoleRenderer';
 import css from './WorkspaceMembers.module.scss';
 import settingsConfig, {
   DEFAULT_COLUMN_WIDTHS,
@@ -103,7 +89,6 @@ const WorkspaceMembers: React.FC<Props> = ({
   pageRef,
   workspace,
 }: Props) => {
-  let { knownRoles } = useStore();
   const rbacEnabled = useFeature().isOn('rbac');
   const { canUpdateRoles } = usePermissions();
   const [canceler] = useState(new AbortController());
@@ -266,90 +251,17 @@ const WorkspaceMembers: React.FC<Props> = ({
       );
     };
 
-    const getAssignedRole = (record: UserOrGroup): V1RoleAssignment | null => {
-      const currentAssignment = assignments.find((aGroup) =>
-        isUser(record)
-          ? !!aGroup?.userRoleAssignments &&
-            !!aGroup.userRoleAssignments.find((a) => a.userId === getIdFromUserOrGroup(record))
-          : !!aGroup?.groupRoleAssignments &&
-            !!aGroup.groupRoleAssignments.find((a) => a.groupId === getIdFromUserOrGroup(record)),
-      );
-      if (isUser(record) && !!record) {
-        if (currentAssignment?.userRoleAssignments) {
-          const myAssignment = currentAssignment.userRoleAssignments.find(
-            (a) => a.userId === getIdFromUserOrGroup(record),
-          );
-          return myAssignment?.roleAssignment || null;
-        }
-      } else if (currentAssignment?.groupRoleAssignments) {
-        const myAssignment = currentAssignment.groupRoleAssignments.find(
-          (a) => a.groupId === getIdFromUserOrGroup(record),
-        );
-        return myAssignment?.roleAssignment || null;
-      }
-      return null;
-    };
-
-    const roleRenderer = (value: string, record: UserOrGroup) => {
-      const roleAssignment = getAssignedRole(record);
-      return (
-        <Select
-          className={css.selectContainer}
-          disabled={!userCanAssignRoles || !roleAssignment?.scopeWorkspaceId}
-          value={roleAssignment?.role?.roleId}
-          onSelect={async (value: RawValueType | LabelInValueType) => {
-            const roleIdValue = value as number;
-            const userOrGroupId = getIdFromUserOrGroup(record);
-            const oldRoleIds = roleAssignment?.role?.roleId ? [roleAssignment?.role?.roleId] : [];
-
-            try {
-              // Try to remove the old role and then add the new role
-              isUser(record)
-                ? await removeRolesFromUser({
-                    roleIds: oldRoleIds,
-                    scopeWorkspaceId: workspace.id,
-                    userId: userOrGroupId,
-                  })
-                : await removeRolesFromGroup({
-                    groupId: userOrGroupId,
-                    roleIds: oldRoleIds,
-                    scopeWorkspaceId: workspace.id,
-                  });
-              try {
-                isUser(record)
-                  ? await assignRolesToUser({
-                      roleIds: [roleIdValue],
-                      scopeWorkspaceId: workspace.id,
-                      userId: userOrGroupId,
-                    })
-                  : await assignRolesToGroup({
-                      groupId: userOrGroupId,
-                      roleIds: [roleIdValue],
-                      scopeWorkspaceId: workspace.id,
-                    });
-              } catch (addRoleError) {
-                handleError(addRoleError, {
-                  publicSubject: 'Unable to update role for user or group unable to add new role.',
-                });
-              }
-            } catch (removeRoleError) {
-              handleError(removeRoleError, {
-                publicSubject:
-                  'Unable to update role for user or group could unable to remove current role.',
-              });
-            }
-          }}>
-          {knownRoles.map((role) => (
-            <Select.Option key={role.id} value={role.id}>
-              {role.name}
-            </Select.Option>
-          ))}
-        </Select>
-      );
-    };
+    const roleRenderer = (value: string, record: UserOrGroup) => (
+      <RoleRenderer
+        assignments={assignments}
+        userCanAssignRoles={userCanAssignRoles}
+        userOrGroup={record}
+        workspaceId={workspace.id}
+      />
+    );
 
     const actionRenderer = (value: string, record: UserOrGroup) => {
-      const assignedRole = getAssignedRole(record);
+      const assignedRole = getAssignedRole(record, assignments);
 
       return userCanAssignRoles && assignedRole?.role.roleId ? (
         <GroupOrMemberActionDropdown
@@ -388,7 +300,7 @@ const WorkspaceMembers: React.FC<Props> = ({
         title: '',
       },
     ] as ColumnDef<UserOrGroup>[];
-  }, [assignments, knownRoles, nameFilterSearch, tableSearchIcon, userCanAssignRoles, workspace]);
+  }, [assignments, nameFilterSearch, tableSearchIcon, userCanAssignRoles, workspace]);
 
   return (
     <div className={css.membersContainer}>
